print.GPArotation {GPArotation}	R Documentation

Print and Summary Methods for GPArotation

Description

Print an object or summary of an object returned by GPFRSorth, GPFRSoblq, GPForth, or GPFoblq.

A GPArotation object by default to will print sorted loadings, Phi, and rotation matrix (if requested).

Output includes contributions of factors SS loadings (Sum of Squared loadings), (see e.g. Harman 1976, sections 2.4 and 12.4).

Usage

    ## S3 method for class 'GPArotation'
print(x, digits=3, sortLoadings=TRUE, rotateMat=FALSE, Table=FALSE, ...)
    ## S3 method for class 'GPArotation'
summary(object, digits=3, Structure=TRUE, ...)
    ## S3 method for class 'summary.GPArotation'
print(x, ...)

Arguments

x
a GPArotation object to summarize.

object
a summary.GPArotation to print.

digits
precision of printed numbers.

sortLoadings
display sorted loadings of a GPArotation object.

rotateMat
display the rotation matrix of a GPArotation object.

Table
display the convergence table of a GPArotation object.

Structure
display the structure matrix of a summary.GPArotation object of a oblique rotation.

...
further arguments passed to other methods.

Details

For examples of print and summary functions, see GPForth.

Value

The object printed or a summary object.

References

Harman, H.H. (1976). Modern Factor Analysis. The University of Chicago Press.

See Also

GPForth, summary

[Package GPArotation version 2025.3-1 Index]